Update every numeric datatype to int and updated the logic accordingly

+++ Battlefield.java
@@ -1,19 +1,19 @@
 public class Battlefield {
     final int width;
     final int height;
-    final byte[][] coordinateSystem; //0 is free, 1 is already shot at, 2 is occupied, 3 is hit //PRÜFEN: int, damit das Spielfeld größer werden kann? Was ist mit Overflow? 
-    private byte occupiedFields;
-    private byte hitFields;
+    final int[][] coordinateSystem; //0 is free, 1 is already shot at, 2 is occupied, 3 is hit //PRÜFEN: int, damit das Spielfeld größer werden kann? Was ist mit Overflow? 
+    private int occupiedFields;
+    private int hitFields;
 
     public Battlefield(int width, int height) {
         this.width = width;
         this.height = height;
-        this.coordinateSystem = new byte[height][width];
+        this.coordinateSystem = new int[height][width];
     }
 
     public boolean setShip(Ship ship) {
-        for(byte i = 0; i < 2; i++) {
-            for(byte position : ship.getPositions()) {
+        for(int i = 0; i < 2; i++) {
+            for(int position : ship.getPositions()) {
                 if(i == 0) {
                     boolean isInBounds = ship.isHorizontal 
                         ? position <= width && position >= 0
@@ -39,9 +39,9 @@ public class Battlefield {
         }
         return true;
     }
-    public boolean hitField(byte x, byte y) {
+    public boolean hitField(int x, int y) {
         try{
-            byte shotField = coordinateSystem[x][y];
+            int shotField = coordinateSystem[x][y];
             if(shotField == 0 || shotField == 2) {
                 coordinateSystem[x][y]++;
                 return true;
